reCAPTCHA fix, moved the logic to a better place, so there is no copy-pasting
This is still the same reCAPTCHA fix (display the CAPTCHA in the right
language), but the logic got moved to it's own class. That way the fix can be
used on all CAPTCHA's instead of copy-pasting the same fix all over again. As
copy-pasting is always a bad idea a better solution was needed.

This doesn't autowire the properties yet though, so there is still some work to
be done!