Added a .editorconfig file
I've added a .editorconfig file, this should help sticking to the code
convention. This could especially be useful when open sourcing the project. If
your editor supports .editorconfig files it will read down the tree, until it
hits a .editorconfig file. This file is in the root and has root = true, so your
editor won't search any further. After having found the file it will set the
right settings for the project.

I hope I've covered all our usual filetypes. I've included .java .css .ftl .js
and .xml
I think we use 4 spaces in all our files, so maybe it would be better to move
the indent_size to the * section.