header.main h1 padding increased to 170px;
I've increaded the padding on the header, because on smaller screens
(widht < 1100px) the word Delivery (h1) started to move over the IISH logo.
By increasing the value to 170px this doesn't happen any more. This only
happened on the Dutch version of the site though, because that logo is a little
longer. I don't think this should have an effect on other things, because it is
a pretty specific selector.