Added overview of existing date exceptions
Added overview html page to show the existing date exceptions.
Removed submit assign from the date_exception html file
Added a link to the overview page on the base html page
Added messages for the exception overview page in both nl and en
Adjusted the isEndDateBeforeBeginDate method so it works properly
Added a get method to pass the existing date exceptions to the overview page.