Updated the reservation exception date
Took all methods regarding the ReservationDateException out of the ReservationController and placed it in the ReservationDateExceptionController. Plus cleaned it up a bit.
Added correct comments to methods in ReservationDateExceptionDAO and ReservationDateExceptionDAOImpl.
Adjusted the toString method of ReservationDateException.
Took methods regarding ReservationDateException out of ReservationService and ReservationServiceImpl and placed them in ReservationDateExceptionService and ReservationDateExceptionImpl and updated some methods.
Added new messages to the messages properties for both "en" and "nl".
Changed the link reference for ReservationDateException to a new link in base.ftl.
Updated the ReservationDateException page to also contain the overview of the already existent ReservationDateExceptions. Threw away the overview page.
Furthermore adjusted the code to get all exceptions to check whether they are before the current date, to only show the future date exceptions in the overview.