Added reminder mail features and little cleanup
Added a reminder mail method to get information from the reproduction and send the mail.
Adjusted the check for cancellation to three work weeks instead of two and added a check for reminder set to two work weeks.
Added the value prop_reproductionMaxDaysReminder to the config.
Added new messages for the reminder.
Adjusted the layout of the standard mail form for the display of the IISG information.
Added a payment reminder mail template with the right message and subject.