Update on Spring Boot conversion
Now displays all pages that are clickable in the header.
.ftl files relocated to corresponding folders.
Added .html to the return calls for the view files.
Language change is now working.
DeliveryProperties now automatically loaded for the view files.
Adjusted the way the deliveryproperties are loaded into the java classes.
Added an empty messages.properties to make sure the messages are working according to Spring Boot documentation.
Removed dependencies no longer used in Spring Boot.