Update
Added spring boot starter mail dependency.
Added systemproperty to handle the encoded slashes in the url.
Added code to let the firewall pass the url with encoded slashes.
Adjusted the calls to the mail templates by adding "mail/" to the return values.
Updated freemarker configuration.
Removed the Waterfilter from the CaptchaEngine due to it not working with Java 1.8