Update for Spring Boot
Added jpamodelgen in pom file for the generation of the abstract classes.
Removed unused properties from DeliveryProperties for these are autoconfigured.
Removed the mail configuration as this is done automatically.
Made sure the access to different methods is handled outside the security configuration by update the SecurityConfiguration and changing @Secured to @PreAuthorize. This done on all controller classes.
Adjusted the UrlDecoder.decode as this was deprecated, had to wrap it with a try catch.
Added error templates for different types of errors along with the messages required for every error. These are just simple static templates.
Furthermore cleaned up some of the code.