Update for spring boot
Cleaned up the POM file of Delivery.
Changed the check for "h2" profile in ConfigConfiguration to "development" for ease of use.
Removed the check for "h2" in SecurityConfiguration and only check on "development" for ease of use.
Added a properties file specifically for development.
Set freemarker.cache explicitly to true for production.