Updated the way prices are given for items other than books/brochures
Added a class Copies, similar to Pages but specifically for copies (not books/brochures).
Added methods to get the number of copies and the price by number of copies to Record.java
Added messages for the copies.
Adjusted the ftl to check whether an item is type BOOk, otherwise it states copies.
Added macro holdingPriceOther to utils.ftl which is similar to holdingPrice to display the price for the order.