Updated the invoice template
Added new messages for the invoice to be displayed on the pdf.
Adjusted the pdf.ftl template to fit for the invoice, this by adding information on the bottom about IISG and adjusting the top of the page to have more of a look and feel of an invoice.
Updated the invoice pdf to be more like an invoice.
Adjusted the display of the name of the orderer.
Added information about the order on the top of the page.
Added a summary below the list of ordered items to display the price and discount and tax. Which calculates an entire price with tax.