Update based on the comments given.
Made a generic method to determine the price for either pages or copies.
Adjusted initReproduction to save the amount of copies/pages if one of both contains an amount, otherwise it is 1.
Simplified the method determineNumberOfCopies for it only has to check for the combination of <amount of pages> <ex>.
Adjusted the holdingPrice macro in mail.ftl.
Updated the calls to the macro holdingPrice from mail.ftl by adding the materialType.
Adjusted the holdingPrice in utils.ftl and removed the holdingPriceOther macro.
Updated the code with calls to the holdingPrice macro from utils.ftl by removing the checks for materialType, since they are done in the macro itself.