Changed the way barcodes are generated and read.
Changed the information given to the constructor for RequestPrintable by adding a parameter for HoldingReproduction and HoldingReservation.
Changed the call to this constructor in multiple classes to add one of these if when available.
Adjusted the way the holding the reservation or reproduction is tied to is being looked for. Now it looks for the reservation or reproduction according to the id, and then gets the holding tied to that reproduction or reservation.