Update how the barcode is generated and retrieved according to feedback
Added getById methods to the HoldingReproductionDAO and the HoldingReservationDAO.
Added a getId() method to HoldingRequest and made HoldingReproduction and HoldingReservation override this method.
Removed HoldingReproduction and HoldingReservation from the RequestPrintable constructor and all the calls to that constructor.
HoldingReservation and HoldingReproduction is now retrieved via HoldingReservationDAO and HoldingReproductionDAO instead of own method.