Adjusted the macro userbase and userHeading
Added a parameter to check whether the language buttons should be disabled. Default value is false, so they will be displayed. By having default value, only specific templates need to call the macro with the correct parameter.
Also set the value disabled for reservation succes to true, so buttons are not displayed.